Move '/' and '/rss' routes into 'main' blueprint
and update templates.
