+++ nyaa/templates/search_results.html
@@ -62,9 +62,9 @@
 				{% set icon_dir = config.SITE_FLAVOR %}
 				<td style="padding:0 4px;">
 				{% if use_elastic %}
-				<a href="{{ url_for('home', c=cat_id) }}" title="{{ category_name(cat_id) }}">
+				<a href="{{ url_for('main.home', c=cat_id) }}" title="{{ category_name(cat_id) }}">
 				{% else %}
-				<a href="{{ url_for('home', c=cat_id) }}" title="{{ torrent.main_category.name }} - {{ torrent.sub_category.name }}">
+				<a href="{{ url_for('main.home', c=cat_id) }}" title="{{ torrent.main_category.name }} - {{ torrent.sub_category.name }}">
 				{% endif %}
 					<img src="{{ url_for('static', filename='img/icons/%s/%s.png'|format(icon_dir, cat_id)) }}" alt="{{ category_name(cat_id) }}">
 				</a>
