+++ nyaa/views/__init__.py
@@ -1,11 +1,13 @@
 from nyaa.views import (
     account,
     admin,
+    main,
     site,
     users,
 )
 
 account_bp = account.bp
 admin_bp = admin.bp
+main_bp = main.bp
 site_bp = site.bp
 users_bp = users.bp
